+++ src/NowoWalletQrBundle.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Nowo\WalletQrBundle;
 
+use LogicException;
+use Nowo\QrCodeBundle\NowoQrCodeBundle;
 use Nowo\WalletQrBundle\DependencyInjection\NowoWalletQrExtension;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
@@ -11,6 +13,8 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 /**
  * Symfony bundle for Google Wallet and Apple Wallet QR save links.
+ *
+ * Requires {@see NowoQrCodeBundle} (nowo-tech/qr-code-bundle) to be registered.
  *
  * @author Héctor Franco Aceituno <hectorfranco@nowo.tech>
  * @copyright 2026 Nowo.tech
@@ -29,5 +33,9 @@ final class NowoWalletQrBundle extends Bundle
     public function build(ContainerBuilder $container): void
     {
         parent::build($container);
+
+        if (!$container->hasExtension('nowo_qr_code')) {
+            throw new LogicException('WalletQrBundle requires NowoQrCodeBundle (nowo-tech/qr-code-bundle). Register Nowo\\QrCodeBundle\\NowoQrCodeBundle in config/bundles.php.');
+        }
     }
 }